requirements.txt
fastapi
uvicorn
requests
PyExecJS